// frontend/src/gen/api/src/models/ESystemNotificationType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ESystemNotificationType = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3
} as const;
export type ESystemNotificationType = typeof ESystemNotificationType[keyof typeof ESystemNotificationType];


export function instanceOfESystemNotificationType(value: any): boolean {
    return Object.values(ESystemNotificationType).includes(value);
}

export function ESystemNotificationTypeFromJSON(json: any): ESystemNotificationType {
    return ESystemNotificationTypeFromJSONTyped(json, false);
}

export function ESystemNotificationTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ESystemNotificationType {
    return json as ESystemNotificationType;
}

export function ESystemNotificationTypeToJSON(value?: ESystemNotificationType | null): any {
    return value as any;
}